feat(matches-reponse): add response for history. fix: fix names misseds

+++ src/main/resources/db/migration/V3__create-table-refresh-token.sql
@@ -2,7 +2,7 @@ create table refresh_tokens(
     id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     token TEXT not null,
     expires_at timestamp not null,
-    user_id integer not null,
+    user_id BIGINT not null,
     foreign key (user_id) references users(id) on delete cascade,
     created_at timestamp not null default current_timestamp,
     updated_at timestamp not null default current_timestamp
